+++ backfit_generator/example/example.dart
@@ -1,4 +1,6 @@
 
+import 'dart:io';
+
 import 'package:backfit/backfit.dart';
 
 import 'my_post.dart';
@@ -9,4 +11,8 @@ part 'example.backfit.dart';
 abstract class Posts {
   @Get('posts')
   Future<Response<MyPost>> postsByUserId(@Query('userId') String userId); 
+
+  @multiPart
+  @Post('photos')
+  Future<Response> uploadFile(@PartFile("image", "media/*") File file);
 }