+++ backfit_generator/lib/src/index.dart
@@ -1,2 +1,2 @@
-export 'generator.dart';
 export 'annotations_processor.dart';
+export 'generator.dart';