+++ backfit_generator/lib/src/processors/class_processor.dart
@@ -1,0 +1,8 @@
+
+import 'package:backfit_generator/src/annotations_processor.dart';
+
+class ClassProcessor {
+  final AnnotationsProcessor _annotationsProcessor;
+
+  ClassProcessor(this._annotationsProcessor);
+}