+++ example/bin/main.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:example/network.dart';
 
 void main(List<String> arguments) {
@@ -5,5 +7,6 @@ void main(List<String> arguments) {
   apiService.postsByUserId('1').catchError((e) {
     print(e.toString());
   });
+
 
 }