+++ example/lib/posts_service.backfit.dart
@@ -35,4 +35,40 @@ mixin _$Posts implements Posts {
       headers: HttpHeaders.fromMap(res.headers),
     );
   }
+
+  @override
+  Future<Response> uploadFile(File file) async {
+    final request =
+        MultipartRequest('POST', Uri.parse('${_client!.baseUrl}/photo'));
+    request.files.add(await MultipartFile.fromPath('image', file.path,
+        contentType: MediaType('media', '*')));
+    final res = await _client!.send(request);
+    final data = await res.stream.bytesToString();
+
+    return Response(
+      data: data.isNotEmpty ? json.decode(data) : null,
+      statusCode: res.statusCode,
+      reasonPhrase: res.reasonPhrase,
+      headers: HttpHeaders.fromMap(res.headers),
+    );
+  }
+
+  @override
+  Future<Response> uploadFiles(List<File> file) async {
+    final request =
+        MultipartRequest('POST', Uri.parse('${_client!.baseUrl}/photos'));
+    for (final _$x in file) {
+      request.files.add(await MultipartFile.fromPath('image', _$x.path,
+          contentType: MediaType('media', '*')));
+    }
+    final res = await _client!.send(request);
+    final data = await res.stream.bytesToString();
+
+    return Response(
+      data: data.isNotEmpty ? json.decode(data) : null,
+      statusCode: res.statusCode,
+      reasonPhrase: res.reasonPhrase,
+      headers: HttpHeaders.fromMap(res.headers),
+    );
+  }
 }