+++ example/lib/posts_service.dart
@@ -1,4 +1,6 @@
 
+import 'dart:io';
+
 import 'package:backfit/backfit.dart';
 import 'my_post.dart';
 
@@ -8,4 +10,14 @@ part 'posts_service.backfit.dart';
 abstract class Posts {
   @Get('posts')
   Future<Response<List<MyPost>>> postsByUserId(@Query('userId') String userId); 
+
+  
+  @multiPart
+  @Post('photo')
+  Future<Response> uploadFile(@PartFile('image', 'media/*') File file);
+
+    
+  @multiPart
+  @Post('photos')
+  Future<Response> uploadFiles(@PartFile('image', 'media/*') List<File> file);
 }