+++ backfit/CHANGELOG.md
@@ -1,3 +1,14 @@
+## [0.0.2+3] 17/09/2021
+
+* Multipart headers support.
+* Fix missing endpoint in multipart request.
+## [0.0.2+2] 16/09/2021
+
+* Added support for deserialization of Multipart request responses.
+* Added PrettyLogger inspired by [pretty_dio_logger](https://pub.dev/packages/pretty_dio_logger)
+## [0.0.2-a] 14/09/2021
+
+* Added preliminary support for Multipart requests.
 ## [0.0.2] 30/08/2021
 
 * Add compatibility with Flutter Web.