+++ backfit_generator/CHANGELOG.md
@@ -1,3 +1,20 @@
+## [0.0.2+5] 10/11/2021
+
+* Http Status Handling 200 - 206.
+## [0.0.2+4] 25/10/2021
+
+* Add support for list of files in multipart requests.
+
+## [0.0.2+3] 17/09/2021
+
+* Fixed missing endpoint and media type from Multipart requests.
+* Fixed multipart request not using backfit client.
+## [0.0.2+2] 16/09/2021
+
+* Added support for deserialization of Multipart request responses.
+## [0.0.2-a] 14/09/2021
+
+* Added preliminary support for Multipart requests.
 ## [0.0.2] 30/08/2021
 
 * Now using PartBuilder instead of Shared Part Builder